added an SRR test script
It takes one random accession from a file to unbar, re-kar, and compare
dump outputs. This is not going to be part of the make file since it is
likely to be moved elsewhere for automation, but for now and for those
testing the new tool, it can be run manually.